Augs, Classes, UI, Sword+Grenade vs structures
2024-09-25
Another very busy week.  Had a bit of a crash for a day or two due to stress and work, and not taking enough time to take care of life.  I was rushing too much for sure.  I’m trying to slow down and relax, catching up on life admin.  Feeling better every day for a few now.
- Made the laser sword not draw it’s collider over walls or structures.
- Also now dynamically adjusting the sword’s trail renderer in an attempt to not let it go through walls.  It’s a little janky up when standing up close to structures, due to no colliders inside the structures, but it works. Could add check for the sword handle being over a structure tile, or going into/out of structures, but I think I’ll leave it for now.
- Did a bunch of work with Augs and their UI panel.  Ironed out a bunch of bugs and got things working nicely.  Need to adjust the size of the panel space in the mothership image, to make more room.
- Added UI elements to death panel to show night reached, previous record, research points gained, and total research points.  Plan to show the math from the difficulty multiplier here too.
- Changed the pause menu buttons so that it only allows you to ‘Give Up’ rather than choose to quit, go to ship, main menu.  This way it shows the rewards for the progress at that point, and the player can still exit quickly.
- Fixed issue with menus not behaving properly while tutorial panels were open with the game paused.  Ended up just allowing control of the menu’s while in the tutorial window.
- Added QCube grenade damage
- Raised platforms have even more HP now.
- Halved grenade damage to structures, QCube, Player.
- Found and fixed bug with turrets not initialized properly with augs.  Was an order of operations issue.
- Added 3 player Classes.  Fighter, Neutral, and Builder.
- Fighter increases structure and structure upgrade costs by 20% but gets 20% off weapon and exosuit upgrades.
- Builder has the opposite of fighter but also gets 1 turret slot at start, and repairs twice as much for the same cost.
- Added the class UI selection panel the augs panel in the mothership scene.  The buttons are simple and boring, I feel they need to be nice images or have an icon of some kind.
- Description text added for Classes.
- Added messages for 10 seconds until night countdown, night started, night survived, and how many turrets are built out of the max when turrets are selected or built.
- Added RP multiplier for difficulty levels
- Did a proper playthrough for the first time in a week since all the changes and I really enjoyed the new features!  They certainly add more dopamine and a drive to play again!  There is much room for improvement here still, which is exciting.
- The player’s look Lerp function needs to be adjusted so that it moved more quickly at the start, and slows down sooner towards the target to allows for snappier precision aiming.  Starting to consider getting rid of rotation speed upgrades and such, and just have it super-fast all the time.
Tomorrow:
Lots of little things that I noticed in my recent playthrough, but nothing major.  I might do more testing and refining, or I might add some other wanted features (probably the better idea).
- Need repair gun upgrades, will probably add them to structures upgrades
-Still need to add pickups for mutant tissues, which are collected by the player and dropped off at the QCube in exchange for credits.  I’ll use gravity for pickups. I’d like to have a slot machine type effect that unloads the tissues one by one into the cube with Credit amounts that pop up with satisfying sound.

+++ Mutant Mayhem v0.1.0-alpha/Assets/_Scripts/Upgrades/Upgrades.cs
@@ -467,7 +467,7 @@ public class MaxTurretsUpgrade : Upgrade
     public override void Apply(StructureStats structureStats, int level)
     {
         Debug.Log("MaxTurretsUpg applied");
-        structureStats.maxTurrets += UpgAmount;
+        structureStats.maxTurrets += (int)UpgAmount;
     }
 
     public override int CalculateCost(Player player, int baseCost, int level)
